HARP-8744 Add custom flat material for shadows
This improves the lighting of flat objects, such that the final rendered pixel
color isn't impacted by a diffuse component. Previously the color would be based
on the direction (using a diffuse shading model), however this dramatically
changes the color of the scene and would require a lot of rework from the
designers / different theme files for shadow / non shadowed themes.

- "fill" techniques can now handle shadows, when shadows are enabled, they now
  use a tweaked version of the standard material.
- Added "shadowIntensity" to the description for directional lights in the
  theme. Note, this value is only useful for the shadows on the
  MapMeshFlatStandardMaterial, because the buildings use the standard three.js
  material, and such a concept doesn't exist there.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>